# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GAC授权中心",
        "title": "GAC授权中心",
        "termsOfService": "https://paymedia.yuque.com/os/hb9uhc",
        "contact": {
            "name": "org.pm.gac.service",
            "url": "https://www.apifox.cn/web/main"
        },
        "license": {
            "name": "GIN.swagger",
            "url": "https://github.com/swaggo/swag/blob/master/README_zh-CN.md#%E5%A6%82%E4%BD%95%E4%B8%8Egin%E9%9B%86%E6%88%90"
        },
        "version": "2.0"
    },
    "host": "api.t.dev.pay.fun",
    "basePath": "/gac/v1",
    "paths": {
...
        "/admin/service_role/create": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GAC/服务商角色管理"
                ],
                "summary": "添加角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.AddRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    }
                }
            }
        },
        "/admin/service_role/delete": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GAC/服务商角色管理"
                ],
                "summary": "角色删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.UpdateRoleStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    }
                }
            }
        },
        "/admin/service_role/disable": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GAC/服务商角色管理"
                ],
                "summary": "角色禁用/启用",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.UpdateRoleStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    }
                }
            }
        },
        "/admin/service_role/page": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GAC/服务商角色管理"
                ],
                "summary": "角色列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.RolePageReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/adminApi.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/assembler.RoleAssembler"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    }
                }
            }
        },
        "/admin/service_role/update": {
            "post": {
                "security": [
                    {
                        "x-token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GAC/服务商角色管理"
                ],
                "summary": "更新角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "description": "The object content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.UpdateRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/adminApi.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "adminApi.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "assembler.RoleAssembler": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "角色代号",
                    "type": "string"
                },
                "created_ip": {
                    "description": "创建Ip",
                    "type": "string"
                },
                "created_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "created_user": {
                    "description": "创建人",
                    "type": "string"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "pk": {
                    "type": "integer"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "1:启用 2：禁用 3：删除",
                    "type": "integer"
                },
                "type": {
                    "description": "角色类别 1:gac管理员 2：sass员工 3：sass服务商",
                    "type": "integer"
                },
                "updated_time": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "req.AddRoleReq": {
            "type": "object",
            "required": [
                "name",
                "status",
                "type"
            ],
            "properties": {
                "code": {
                    "description": "角色code",
                    "type": "string"
                },
                "created_ip": {
                    "description": "创建者Ip",
                    "type": "string"
                },
                "created_user": {
                    "description": "创建人",
                    "type": "string"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "description": "状态：1:启用 2：停用  3:  删除",
                    "type": "integer"
                },
                "type": {
                    "description": "角色类别 1:管理员 2：普通用户 3： 服务商",
                    "type": "integer"
                }
            }
        },
...
        "req.RolePageReq": {
            "type": "object",
            "properties": {
                "created_user": {
                    "description": "创建人名称",
                    "type": "string"
                },
                "current": {
                    "description": "当前页",
                    "type": "integer"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "page_size": {
                    "description": "页大小",
                    "type": "integer"
                },
                "pagination": {
                    "description": "是否使用分页查询",
                    "type": "boolean"
                },
                "pk": {
                    "description": "pk",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：1:启用 2：停用  3:  删除",
                    "type": "integer"
                },
                "type": {
                    "description": "角色类别 1:管理员 2：普通用户 3： 服务商",
                    "type": "integer"
                }
            }
        },
...
        "req.UpdateRoleReq": {
            "type": "object",
            "required": [
                "name",
                "status",
                "type"
            ],
            "properties": {
                "code": {
                    "description": "角色code",
                    "type": "string"
                },
                "created_ip": {
                    "description": "创建者Ip",
                    "type": "string"
                },
                "created_user": {
                    "description": "创建人",
                    "type": "string"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "pk": {
                    "description": "Pk",
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "description": "状态：1:启用 2：停用  3:  删除",
                    "type": "integer"
                },
                "type": {
                    "description": "角色类别 1:管理员 2：普通用户 3： 服务商",
                    "type": "integer"
                }
            }
        },
        "req.UpdateRoleStatusReq": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "created_ip": {
                    "description": "创建者Ip",
                    "type": "string"
                },
                "created_user": {
                    "description": "创建人",
                    "type": "string"
                },
                "pk": {
                    "description": "Pk",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：1:启用 2：停用",
                    "type": "integer"
                }
            }
        },
...
    }
}